Re-organise pep_processor
