+++ pep_extensions/__init__.py
@@ -4,9 +4,9 @@ from sphinx.application import Sphinx
 from docutils.writers.html5_polyglot import HTMLTranslator
 
 from pep_extensions.config import __version__
-from pep_extensions.pep_processor import pep_html_translator
-from pep_extensions.pep_processor import pep_parser
-from pep_extensions.pep_processor import pep_role
+from pep_extensions.pep_processor.html import pep_html_translator
+from pep_extensions.pep_processor.parsing import pep_parser
+from pep_extensions.pep_processor.parsing import pep_role
 from pep_extensions.pepzero.generate_pep_index import create_pep_zero
