+++ pep_extensions/pep_processor/html/pep_html_translator.py
@@ -1,8 +1,7 @@
 from docutils import nodes
 import sphinx.writers.html5 as html5
 import re
-import logging
-logger = logging.getLogger("sphinx")
+
 
 class PEPTranslator(html5.HTML5Translator):
     compact_field_list = True
@@ -10,7 +9,6 @@ class PEPTranslator(html5.HTML5Translator):
 
     def __init__(self, *args):
         super(PEPTranslator, self).__init__(*args)
-        logger.warning("PEP Translator")
         self.compact_simple: bool = False
 
     # Omit <p> tags to produce visually compact lists
